Added more the the Math submodule and created Misc submodule
Created the Misc submodule to hold the random things from the original library.
Updated the Math library to include a few more functions; remainder, squared,
and s_square_root.

There were more functions in the original library, but implementing them
involves breaking a lot of look and feel of the library. This will
likely be the final state of the library for a bit, barring minor
addtional changes.